Fix critical bug in BF2 update procedure. Perform timer callback functionality only upon timeout
When using a steady timer, TimerCallback is being called not only when a timeout occurs, but also after the timer was aborted.
The TimerCallback function is only necessary when the timer has expired.
Therefore, this commit adds a query to verify that the callback is performed only upon timeout.

The fix was tested using dbus and the Redfish interface. Upon timeout, the objects are being updated properly (Activation:Failed, Progress:30) and Redfish task get exception as expected.

Fixes nvbug https://redmine.mellanox.com/issues/3807543